requirements.txt
plotly>=4.8.1
Pillow>=6.0.0
matplotlib>=3.1.0
numpy>=1.16.4
pandas>=0.24.2
scikit-learn>=0.21.3
scipy>=1.3.0
seaborn>=0.9.0
tensorboardX>=1.7
torch>=1.7.1
torchsummary>=1.5.1
torchvision>=0.8.2
umap-learn>=0.3.9
